+++ Data/Dungeons.lua
@@ -26,7 +26,7 @@ TF_DUNGEONS = {
     },
     { zone = TF_ZONE.THE_BARRENS, x = 0.493, y = 0.93, name = "Razorfen Downs", minLevel = 35, recommendedLevel = { 37, 46 }, entrance = { zone = TF_ZONE.KALIMDOR, x = 0.5325, y = 0.7117 } },
     { zone = TF_ZONE.BADLANDS, x = 0.4453, y = 0.1212, name = "Uldaman", minLevel = 30, recommendedLevel = { 41, 51 }, entrance = { zone = TF_ZONE.EASTERN_KINGDOMS, x = 0.5385, y = 0.5764 } },
-    { zone = TF_ZONE.TANARIS, x = 0.3872, y = 0.2, name = "Zul'Farrak", minLevel = 39, recommendedLevel = { 42, 46 } },
+    { zone = TF_ZONE.TANARIS, x = 0.3872, y = 0.2, name = "Zul'Farrak", minLevel = 39, recommendedLevel = { 44, 54 } },
     { zone = TF_ZONE.DESOLACE, x = 0.291, y = 0.625, name = "Maraudon", minLevel = 30, recommendedLevel = { 46, 55 } },
     { zone = TF_ZONE.SWAMP_OF_SORROWS, x = 0.70, y = 0.53, name = "The Temple of Atal'Hakkar", minLevel = 38, recommendedLevel = { 50, 56 }, entrance = { zone = TF_ZONE.EASTERN_KINGDOMS, x = 0.5681, y = 0.7515 } },
     {