+++ changelog.md
@@ -1,6 +1,7 @@
 ### v1.0.9 (Aug. 2x, 2020)
 * Added Searing Gorge NPCs.
 * Added Eastern Plaguelands NPCs.
+* Added Un'Goro Crater NPCs.
 * Updated some instance entrances.
   * Added required key to Scarlet Monastery Armory and Cathedral.
   * Added internal entrance location to Razorfen Downs.
@@ -10,6 +11,7 @@
   * Increased precision of the Stratholme and Naxxramas entrance location.
   * Increased precision of the Uldaman entrance location and added the internal entrance location.
   * Increased precision of the Blackrock Mountain entrance locations, added the internal instance locations, and added the attunement requirements.
+  * Added Drakefire Amulet requirement to Onyxia's Lair.
 * Added a couple NPCs outside of Scholomance.
 
 [See past changes here.](https://github.com/Mokulu/Townsfolk-Tracker/releases)